+++ src/DraggableList/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import DraggableFlatList from 'react-native-draggable-flatlist';
-import type { Props } from './types';
+import type { DraggableListProps } from './types';
 
-export const MyDraggableList = ({ items, ...props }: Props) => {
+export const DraggableList = ({ items, ...props }: DraggableListProps) => {
   return (
     <DraggableFlatList
       data={items}
