+++ src/decorators/ScaleDecorator/index.tsx
@@ -2,6 +2,8 @@ import React from 'react';
 import { ScaleDecorator } from 'react-native-draggable-flatlist';
 import type { Props } from './types';
 
-export function MyScaleDecorator({ children }: Props) {
+function CustomScaleDecorator({ children }: Props) {
   return <ScaleDecorator>{children}</ScaleDecorator>;
 }
+
+export { CustomScaleDecorator as ScaleDecorator };
