+++ src/decorators/ScaleDecorator/types.tsx
